got simd offsets working, minus 2 important edge cases

+++ pco/src/ans/mod.rs
@@ -6,12 +6,12 @@ mod decoding;
 mod encoding;
 mod spec;
 
-// must be u16 or larger (though u32+ may become a requirement in the future)
+// must be u16 or larger
 // should not be exposed in public API
 pub(crate) type AnsState = u32;
 // must be u16 or larger
 // should not be exposed in public API
-pub(crate) type Token = u32;
+pub(crate) type Token = u16;
 
 #[cfg(test)]
 mod tests {
